+++ pkg/network/cache/types.go
@@ -5,24 +5,7 @@ import (
 	"net"
 
 	"github.com/vishvananda/netlink"
-
-	v1 "kubevirt.io/api/core/v1"
 )
-
-type PodIfaceState int
-
-const (
-	PodIfaceNetworkPreparationPending PodIfaceState = iota
-	PodIfaceNetworkPreparationStarted
-	PodIfaceNetworkPreparationFinished
-)
-
-type PodCacheInterface struct {
-	Iface  *v1.Interface `json:"iface,omitempty"`
-	PodIP  string        `json:"podIP,omitempty"`
-	PodIPs []string      `json:"podIPs,omitempty"`
-	State  PodIfaceState `json:"networkState,omitempty"`
-}
 
 type DHCPConfig struct {
 	Name                string
